fix: 引数の形を複数からinputにまとめるように修正

+++ backend/todo_app/schema.py
@@ -1,9 +1,14 @@
 import graphene
 from graphene_django import DjangoObjectType
 from todo_app.models import Task
+from .common import CreateTaskInput, UpdateTaskInput
 
 
 class TaskType(DjangoObjectType):
+    """
+    Taskモデルの全てのフィールドが含まれたGraphQLタイプ
+    """
+
     class Meta:
         model = Task
         fields = "__all__"
@@ -14,72 +19,107 @@ class RootQuery(graphene.ObjectType):
     全てのクエリをまとめるためのGraphQLオブジェクトタイプ
     """
 
-    # 全てのタスクを取得するためのクエリ
-    all_tasks = graphene.List(TaskType, description="全てのタスクを取得する")
+    all_tasks = graphene.List(TaskType, description="タスクの一覧を取得するフィールド")
+    task = graphene.Field(TaskType, description="特定のタスクを取得するフィールド", id=graphene.Int())
 
-    # 特定のタスクを取得するためのクエリ
-    task = graphene.Field(TaskType, description="特定のタスクを取得する", id=graphene.Int())
-
+    # all_tasksフィールドのリゾルバ
     def resolve_all_tasks(self, info):
+        # リクエストからユーザー情報を取得
+        # user = info.context.user
+        # ユーザーが認証されていない場合、例外を投げる
+        # if not user.is_authenticated:
+        # raise Exception("タスクの一覧を取得するにはユーザー認証が必要です。")
+        # 認証されたユーザーのタスクをフィルタリングして返す
+        # return Task.objects.filter(user=user)
         return Task.objects.all()
 
+    # taskフィールドのリゾルバ
     def resolve_task(self, info, id):
-        return Task.objects.get(pk=id)
+        # リクエストからユーザー情報を取得
+        user = info.context.user
+        # ユーザーが認証されていない場合、例外を投げる
+        if not user.is_authenticated:
+            raise Exception("タスクを取得するにはユーザー認証が必要です。")
+
+        # タスクを取得する
+        try:
+            # 与えられたIDと認証されたユーザーに基づいてタスクを取得
+            task = Task.objects.get(pk=id, user=user)
+        # タスクが存在しない場合、例外を投げる
+        except Task.DoesNotExist:
+            raise Exception("タスクが見つかりません。")
+
+        # タスクを返す
+        return task
 
 
-# 新しいタスクを作成するためのGraphQLミューテーション
-class CreateTaskMutation(graphene.Mutation):
+class CreateTask(graphene.Mutation):
+    """
+    新しいタスクを作成するためのGraphQLミューテーション
+    """
+
     # TaskTypeの戻り値の型を定義
     task = graphene.Field(TaskType)
 
     # 引数の設定
     class Arguments:
-        title = graphene.String(required=True)
-        description = graphene.String()
-        completed = graphene.Boolean()
+        input = CreateTaskInput(required=True)
 
     # 新しいタスクを作成してデータベースに保存する
-    def mutate(self, info, title, description=None, completed=False):
+    def mutate(self, info, input):
         # タスクをデータベースに保存し、その内容を変数に格納
         task = Task.objects.create(
-            title=title, description=description, completed=completed
+            title=input.title, description=input.description, completed=input.completed
         )
         # 新しく作成されたタスクのデータを含むミューテーションの結果をクライアントに返す
-        return CreateTaskMutation(task=task)
+        return CreateTask(task=task)
 
 
-# 既存のタスクを更新するためのGraphQLミューテーション
-class UpdateTaskMutation(graphene.Mutation):
-    # TaskTypeの戻り値の型を定義
+class UpdateTask(graphene.Mutation):
+    """
+    既存のタスクを更新するためのGraphQLミューテーション
+    """
+
+    # 更新されたタスクを返すためのフィールド
     task = graphene.Field(TaskType)
 
     # 引数の設定
     class Arguments:
-        id = graphene.ID(required=True)
-        # 更新時に必須項目にしない項目はオプションにする
-        title = graphene.String()
-        description = graphene.String()
-        completed = graphene.Boolean()
+        input = UpdateTaskInput(required=True)
 
-    # 既存のタスクを更新してデータベースに保存する
-    # オプションの引数にはNoneを設定する
-    def mutate(self, info, id, title=None, description=None, completed=None):
-        # データベースからidと一致するpkを持つタスクを取得
-        task = Task.objects.get(pk=id)
-        if title:
-            task.title = title
-        if description:
-            task.description = description
-        if completed is not None:
-            task.completed = completed
-        # データベースのタスクを更新
+    # 既存のタスクを更新してデータベースに保存
+    # オプションの引数にはNoneを設定
+    def mutate(self, info, input):
+        user = info.context.user  # リクエストからユーザー情報を取得
+        if not user.id_authenticated:  # ユーザーが認証されていない場合、例外を投げる
+            raise Exception("タスクを更新するにはユーザー認証が必要です。")
+        try:
+            # IDとユーザーに基づいてタスクを取得
+            task = Task.objects.get(pk=id, user=user)
+        # タスクが存在しない場合、例外を投げる
+        except Task.DoesNotExist:
+            raise Exception("タスクが見つかりません。")
+
+        # 入力引数に応じてタスクのプロパティを更新
+        if input.title:
+            task.title = input.title
+        if input.description:
+            task.description = input.description
+        if input.completed is not None:
+            task.completed = input.completed
+
+        # タスクの変更をデータベースに保存
         task.save()
+
         # 更新されたタスクのデータを含むミューテーションの結果をクライアントに返す
-        return UpdateTaskMutation(task=task)
+        return UpdateTask(task=task)
 
 
-# タスクを削除するためのGraphQLミューテーション
-class DeleteTaskMutation(graphene.Mutation):
+class DeleteTask(graphene.Mutation):
+    """
+    タスクを削除するためのGraphQLミューテーション
+    """
+
     # タスクの削除が成功したかどうかをクライアントに伝えるためのフラグを設定
     ok = graphene.Boolean()
 
@@ -88,22 +128,30 @@ class DeleteTaskMutation(graphene.Mutation):
 
     # 既存のタスクをデータベースから削除する
     def mutate(self, info, id):
-        # データベースからidと一致するpkを持つタスクを取得
-        task = Task.objects.get(pk=id)
+        user = info.context.user
+
+        if not user.is_authenticated:
+            raise Exception("タスクを削除するにはユーザー認証が必要です。")
+        try:
+            task = Task.objects.get(pk=id, user=user)
+        except Task.DoesNotExist:
+            raise Exception("タスクが見つかりません。")
+
         # データベースからタスクを削除
         task.delete()
-        return DeleteTaskMutation(ok=True)
+
+        # タスクが正常に削除されたことをTrueとしてクライアントに返す
+        return DeleteTask(ok=True)
 
 
-# 全てのミューテーションをまとめるためのGraphQLオブジェクトタイプ
 class RootMutation(graphene.ObjectType):
     """
     全てのミューテーションをまとめるためのGraphQLオブジェクトタイプ
     """
 
-    create_task = CreateTaskMutation.Field()
-    update_task = UpdateTaskMutation.Field()
-    delete_task = DeleteTaskMutation.Field()
+    create_task = CreateTask.Field()
+    update_task = UpdateTask.Field()
+    delete_task = DeleteTask.Field()
 
 
 # スキーマの定義: クエリとミューテーションを含むGraphQLスキーマ
